requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
certifi==2023.5.7
charset-normalizer==2.0.12
decorator==5.1.1
Django==4.2.1
idna==3.4
MouseInfo==0.1.3
Pillow==9.5.0
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
python-decouple==3.8
python-dotenv==1.0.0
python-xlib==0.33
python3-xlib==0.15
pytweening==1.0.5
requests==2.26.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urlbox==1.0.6
urllib3==1.26.15
validators==0.18.2
